pyproject.toml
[tool.poetry]
name = "portal"
version = "0.1.0"
description = ""
authors = ["occy88 <octavio.delser@gmail.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.dependencies]
applicationinsights = "*"
arrow = "*"
attrs = "*"
boto3 = "*"
cryptography = "*"
debugpy = "*"
dj-static = "*"
django = "4.2.1"
django-crum = "*"
django-extensions = "*"
django-filter = "*"
django-model-utils = "*"
django-redis = "*"
django-render-block = "*"
django-replicated = "*"
django-taggit = "*"
djangorestframework = "3.14.0"
djangorestframework-camel-case = "*"
drf-jwt = "*"
drf-nested-routers = "*"
drf-partial-response = "*"
drf-spectacular = "*"
elasticsearch = ">=7.0.0,<=7.12.0"
flower = "*"
gunicorn = "*"
ipython = "*"
psycopg2 = "*"
pyclamd = "*"
python = ">=3.8,<3.11"
python-json-logger = "*"
pyyaml = "*"
requests = "*"
sentry-sdk = "1.22.2"
transitions = "*"
urllib3= ">=1.26.11"
model-bakery = "^1.12.0"
djangorestframework-simplejwt = "^5.2.2"
coverage = "^6.5.0"
dj-database-url = "^1.0.0"
django-appconf = "^1.0.5"
django-compressor = "^4.1"
django-compressor-toolkit = "^0.6.0"
django-cors-headers = "^3.13.0"
django-guardian = "^2.4.0"
django-pipeline = "^2.0.8"
django-pipeline-browserify = "^0.6.2"
django-webpack-loader = "*"
drf-yasg = "^1.21.4"
fakeredis = "^2.0.0"
mixer = "^7.2.2"
numpy = "^1.23.5"
pillow = "^9.3.0"
pre-commit = "^2.20.0"
pydot = "^1.4.2"
pyjwt = {extras = ["crypto"], version = "^2.6.0"}
pyparsing = "^3.0.9"
whitenoise = "^6.4.0"
django-ltree = "^0.5.3"
django-storages = {extras = ["azure"], version = "^1.13.2"}
aif360 = {extras = ["lawschoolgpa"], version = "^0.5.0"}
blackboxauditing = "^0.1.54"
captum = "^0.5.0"
celery = "^5.2.7"
dice-ml = "^0.9"
django-slowtests = "^1.1.1"
eli5 = "^0.13.0"
fairlearn = "^0.8.0"
imbalanced-learn = "^0.9.1"
joblib = "^1.2.0"
lime = "^0.2.0.1"
opencv-python-headless = "^4.6.0.66"
pandas = "^1.5.2"
redis = "^4.3.5"
responses = "^0.22.0"
scikit-learn = "^1.1.3"
shap = "^0.41.0"
tblib = "^1.7.0"
tf-explain = "^0.3.1"
torch = "^1.13.1"
torchray = "^1.0.0"
typeguard = "^2.13.3"
xgboost = "^1.7.1"
zstandard = "^0.19.0"
importlib-metadata = "^5.1.0"
tensorflow = { version = "2.11.0", markers = "platform_machine=='x86_64' or sys_platform=='win32'" }
tensorflow-macos = { version = "^2.11.0", markers = "platform_machine!='x86_64' and sys_platform=='darwin'" }
lightgbm = "^3.3.5"
numba = "^0.56.4"
wheel = "^0.38.4"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"
pytest-xdist = "^3.2.0"
pytest-mock = "^3.10.0"
decorator = "^5.1.1"
pytest = "^7.2.1"
dill = "^0.3.6"

[tool.poetry.group.dev.dependencies]
black = "*"
deepdiff = "*"
django-debug-toolbar = "*"
django-debug-toolbar-force = "*"
django-silk = "*"
flake8 = "*"
flake8-bugbear = "*"
isort = "*"
locust = "*"
model_bakery = "*"
pre-commit = "*"
pytest-cov = "*"
pytest-django = "*"
pytest-mock = "*"
pytest-repeat = "*"
pytest-xdist = "*"
requests-mock = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
